+++ ChangeLog.rst
@@ -9,6 +9,7 @@ ChangeLog
 - Use env vars to expose grocker meta-data to the application
 - Expose some meta-data using image labels
 - Use docker build args to pass some build parameters
+- Add application venv bin in PATH
 
 4.1 (2016-09-19)
 ----------------